refactor: move mda components
