+++ src/pymmcore_widgets/__init__.py
@@ -14,14 +14,13 @@ from ._group_preset_widget._group_preset_table_widget import GroupPresetTableWid
 from ._image_widget import ImagePreview
 from ._live_button_widget import LiveButton
 from ._load_system_cfg_widget import ConfigurationWidget
-from ._mda_widget._mda_widget import MDAWidget
+from ._mda import MDAWidget, SampleExplorerWidget
 from ._objective_widget import ObjectivesWidget
 from ._pixel_size_widget import PixelSizeWidget
 from ._presets_widget import PresetsWidget
 from ._properties_widget import PropertiesWidget
 from ._property_browser import PropertyBrowser
 from ._property_widget import PropertyWidget
-from ._sample_explorer_widget._sample_explorer_widget import SampleExplorerWidget
 from ._shutter_widget import ShuttersWidget
 from ._slider_dialog_widget import SliderDialog
 from ._snap_button_widget import SnapButton
