+++ src/pymmcore_widgets/_mda/_channel_table_widget.py
