+++ src/pymmcore_widgets/_mda/_grid_widget.py
