+++ src/pymmcore_widgets/_mda/_sample_explorer_gui.py
@@ -15,14 +15,10 @@ from qtpy.QtWidgets import (
 )
 from superqt import QCollapsible
 
-from pymmcore_widgets._channel_table_widget import ChannelTable
-from pymmcore_widgets._general_mda_widgets import (
-    _MDAControlButtons,
-    _MDAPositionTable,
-    _MDATimeLabel,
-)
-from pymmcore_widgets._time_plan_widget import TimePlanWidget
-from pymmcore_widgets._zstack_widget import ZStackWidget
+from ._channel_table_widget import ChannelTable
+from ._general_mda_widgets import _MDAControlButtons, _MDAPositionTable, _MDATimeLabel
+from ._time_plan_widget import TimePlanWidget
+from ._zstack_widget import ZStackWidget
 
 LBL_SIZEPOLICY = QSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed)
