+++ src/pymmcore_widgets/_mda/_sample_explorer_widget.py
